refactor(api): relocate REST v1 deprecation-header util (TD-V1SUNSET-1 step 2) (#825)
* refactor(api): relocate REST v1 deprecation-header util out of rest::v1 (TD-V1SUNSET-1 step 2)

Prerequisite for hard-deleting the deprecated `crates/platform/proximadb-api/src/rest/v1/` adapter dir. The deprecation-header family (`add_rest_v1_deprecation_headers`, `with_v1_compatibility_headers`, their `apply_v1_deprecation_headers` helper, and the two consts) lived in `proximadb_api::rest::v1::mod`, but `add_rest_v1_deprecation_headers` is imported by LIVE root-crate code (`src/network/rest/v1/handlers.rs`, which serves /api/v2 despite the v1 path segment) — deleting `rest::v1` would break that import.

Move the whole family to a new neutral `proximadb_api::rest::deprecation` module; re-export back into `rest::v1` so the doomed dir's internal `super::with_v1_compatibility_headers` call sites compile unchanged; repoint `handlers.rs` at `rest::deprecation`. Additive, non-breaking. Step 3 (delete `rest::v1/` + drop its re-exports) follows in a separate commit.

* docs(td): reconcile TD-V1SUNSET-1 — step 3 invalidated, step 4 already done

Fresh origin/develop worktree audit disproved the deletion premise:

- Step 4 (gRPC v1 api-crate delete): ALREADY COMPLETE on develop — grpc/v1/ (11 files), grpc/builder.rs, and the lib.rs grpc::v1 re-export are all gone; grpc.rs:5-8 records the removal.

- Step 3 (delete api-crate rest/v1/): INVALIDATED — that dir holds LIVE /api/v2 port-backed routers (create_{graph,document,observability,multimodal,hybrid_search}_router) wired into production at src/network/rest/v1/handlers.rs:1460-1542, NOT deprecated v1 adapters. A blanket git rm -r would fail to compile and remove /api/v2 routes. A module rename is TD-104, not v1-sunset.

- Step 2 (relocate deprecation-header util): LANDED (prior commit b04410b66).

- Step 5 (remove v1_sunset middleware): unchanged, gated on sunset-window verification; sole outstanding REST-side action.

Net: the REST v1 sunset is effectively complete — legacy /api/v1 routes were removed in the API standardization hard-rename and step 1 410s residual traffic; no deprecated route handlers remain, only the misnamed live rest::v1 module dir.
